+++ src/components/SideBar/SideBar.vue
@@ -1,10 +1,14 @@
 <template>
-    <aside class="uf-sidebar uk-visible@s">
+    <aside class="uf-sidebar uk-visible@m" v-bind="$attrs">
         <ul class="uk-nav-default" uk-nav>
             <slot></slot>
         </ul>
     </aside>
-    <div class="uk-hidden@s uk-offcanvas" id="offcanvas-sidebar" uk-offcanvas="overlay: true; selClose: .uf-sidebar-close">
+    <div
+        class="uk-hidden@m uk-offcanvas"
+        id="offcanvas-sidebar"
+        uk-offcanvas="overlay: true; selClose: .uf-sidebar-close"
+        v-bind="$attrs">
         <div class="uf-sidebar uk-offcanvas-bar">
             <button class="uk-offcanvas-close" type="button" uk-close></button>
             <ul class="uk-nav-default" uk-nav>