+++ src/components/SideBar/SideBarLabel.vue
@@ -5,5 +5,5 @@ defineProps<{
 </script>
 
 <template>
-    <li class="uk-nav-header">{{ label }}</li>
+    <li class="uk-nav-header" v-bind="$attrs">{{ label }}</li>
 </template>